+++ src/app/concepts/page.tsx
@@ -1,0 +1,5 @@
+import ConceptsContent from "./ConceptsContent";
+
+export default function ConceptsPage() {
+  return <ConceptsContent />;
+}